+++ src/main/steps/applicant1/hub-page/post.test.ts
@@ -3,9 +3,10 @@ import { mockResponse } from '../../../../test/unit/utils/mockResponse';
 import {
   APPLICANT_1_CONFIRM_RECEIPT,
   APPLICANT_2_CONFIRM_RECEIPT,
+  CITIZEN_APPLICANT2_UPDATE,
+  CITIZEN_UPDATE,
   DRAFT_CONDITIONAL_ORDER,
   State,
-  UPDATE_CONDITIONAL_ORDER,
   YesOrNo,
 } from '../../../app/case/definition';
 import { FormContent } from '../../../app/form/Form';
@@ -73,7 +74,7 @@ describe('HubPagePostController', () => {
     expect(req.locals.api.triggerEvent).toHaveBeenCalledWith('1234', body, DRAFT_CONDITIONAL_ORDER);
   });
 
-  it('triggers UPDATE_CONDITIONAL_ORDER in ConditionalOrderDrafted state', async () => {
+  it('triggers CITIZEN_UPDATE in ConditionalOrderDrafted state for Applicant 1', async () => {
     const body = {
       state: State.ConditionalOrderDrafted,
       applicant1ConfirmReceipt: YesOrNo.YES,
@@ -90,6 +91,26 @@ describe('HubPagePostController', () => {
     const res = mockResponse();
     await hubPagePostController.post(req, res);
 
-    expect(req.locals.api.triggerEvent).toHaveBeenCalledWith('1234', body, UPDATE_CONDITIONAL_ORDER);
+    expect(req.locals.api.triggerEvent).toHaveBeenCalledWith('1234', body, CITIZEN_UPDATE);
+  });
+
+  it('triggers CITIZEN_APPLICANT2_UPDATE in ConditionalOrderDrafted state for Applicant 2', async () => {
+    const body = {
+      state: State.ConditionalOrderDrafted,
+      applicant1ConfirmReceipt: YesOrNo.YES,
+    };
+    const mockFormContent = {
+      fields: {
+        applicant1ConfirmReceipt: {},
+        state: {},
+      },
+    } as unknown as FormContent;
+    const hubPagePostController = new HubPagePostController(mockFormContent.fields);
+
+    const req = mockRequest({ body, session: { isApplicant2: true } });
+    const res = mockResponse();
+    await hubPagePostController.post(req, res);
+
+    expect(req.locals.api.triggerEvent).toHaveBeenCalledWith('1234', body, CITIZEN_APPLICANT2_UPDATE);
   });
 });
