+++ src/main/steps/applicant1/hub-page/post.ts
@@ -3,9 +3,10 @@ import autobind from 'autobind-decorator';
 import {
   APPLICANT_1_CONFIRM_RECEIPT,
   APPLICANT_2_CONFIRM_RECEIPT,
+  CITIZEN_APPLICANT2_UPDATE,
+  CITIZEN_UPDATE,
   DRAFT_CONDITIONAL_ORDER,
   State,
-  UPDATE_CONDITIONAL_ORDER,
 } from '../../../app/case/definition';
 import { AppRequest } from '../../../app/controller/AppRequest';
 import { AnyObject, PostController } from '../../../app/controller/PostController';
@@ -18,7 +19,7 @@ export default class HubPagePostController extends PostController<AnyObject> {
     } else if (req.session.userCase.state === State.AwaitingConditionalOrder) {
       return DRAFT_CONDITIONAL_ORDER;
     } else {
-      return UPDATE_CONDITIONAL_ORDER;
+      return req.session.isApplicant2 ? CITIZEN_APPLICANT2_UPDATE : CITIZEN_UPDATE;
     }
   }
 }
